+++ spec/grape/dsl/inside_route_spec.rb
@@ -117,6 +117,16 @@ module Grape
           expect { subject.status :foo_bar }
             .to raise_error(ArgumentError, 'Status code :foo_bar is invalid.')
         end
+
+        it 'raises error if unknow status code is passed' do
+          expect { subject.status 210 }
+            .to raise_error(ArgumentError, 'Status code 210 is invalid.')
+        end
+
+        it 'raises error if status is not a fixnum or symbol' do
+          expect { subject.status Object.new }
+            .to raise_error(ArgumentError, 'Status code must be Fixnum or Symbol.')
+        end
       end
 
       describe '#header' do